Support timezones-ish
Added a timezone parameter to request
in javascript. If not set, defaults to
utc timezone. Gets the days since last win
by diffing the time at this timezone and
the time gotten from espn, and then making
sure it's more than 0. This doesn't work
perfectly if your timezone is ahead of espn
but at least it won't say -1 days since last
win.